Multiple Regression with Multivariate Adaptive Shrinkage from summary data.

Performs multivariate multiple regression with mixture-of-normals prior.

mr.mash.rss(
  Bhat,
  Shat,
  Z,
  R,
  covY,
  n,
  S0,
  w0 = rep(1/(length(S0)), length(S0)),
  V = NULL,
  mu1_init = NULL,
  tol = 1e-04,
  convergence_criterion = c("mu1", "ELBO"),
  max_iter = 5000,
  update_w0 = TRUE,
  update_w0_method = "EM",
  w0_threshold = 0,
  compute_ELBO = TRUE,
  standardize = TRUE,
  verbose = TRUE,
  update_V = FALSE,
  update_V_method = c("full", "diagonal"),
  version = c("Rcpp", "R"),
  e = 1e-08,
  ca_update_order = c("consecutive", "decreasing_logBF", "increasing_logBF", "random"),
  X_colmeans = NULL,
  Y_colmeans = NULL,
  check_R = TRUE,
  R_tol = 1e-08,
  nthreads = as.integer(NA)
)

Arguments

Bhat

p x r matrix of regression coefficients from univariate simple linear regression.

Shat

p x r matrix of standard errors of the regression coefficients from univariate simple linear regression.

Z

p x r matrix of Z-scores from univariate simple linear regression.

R

p x p correlation matrix among the variables.

covY

r x r covariance matrix across responses.

n

scalar indicating the sample size.

S0

List of length K containing the desired r x r prior covariance matrices on the regression coefficients.

w0

K-vector with prior mixture weights, each associated with the respective covariance matrix in S0.

V

r x r residual covariance matrix.

mu1_init

p x r matrix of initial estimates of the posterior mean regression coefficients. These should be on the same scale as the X provided. If standardize=TRUE, mu1_init will be scaled appropriately after standardizing X.

tol

Convergence tolerance.

convergence_criterion

Criterion to use for convergence check.

max_iter

Maximum number of iterations for the optimization algorithm.

update_w0

If TRUE, prior weights are updated.

update_w0_method

Method to update prior weights. Only EM is currently supported.

w0_threshold

Drop mixture components with weight less than this value. Components are dropped at each iteration after 15 initial iterations. This is done to prevent from dropping some poetentially important components prematurely.

compute_ELBO

If TRUE, ELBO is computed.

standardize

If TRUE, X is "standardized" using the sample means and sample standard deviations. Standardizing X allows a faster implementation, but the prior has a different interpretation. Coefficients and covariances are returned on the original scale.

verbose

If TRUE, some information about the algorithm's process is printed at each iteration.

update_V

if TRUE, residual covariance is updated.

update_V_method

Method to update residual covariance. So far, "full" and "diagonal" are supported. If update_V=TRUE and V is not provided by the user, this option will determine how V is computed (and fixed) internally from mu1_init.

version

Whether to use R or C++ code to perform the coordinate ascent updates.

e

A small number to add to the diagonal elements of the prior matrices to improve numerical stability of the updates.

ca_update_order

The order with which coordinates are updated. So far, "consecutive", "decreasing_logBF", "increasing_logBF", "random" are supported.

X_colmeans

a p-vector of variable means.

Y_colmeans

a r-vector of response means.

check_R

If TRUE, R is checked to be positive semidefinite.

R_tol

tolerance to declare positive semi-definiteness of R.

nthreads

Number of RcppParallel threads to use for the updates. When nthreads is NA, the default number of threads is used; see defaultNumThreads. This setting is ignored when version = "R".

Value

A mr.mash.rss fit, stored as a list with some or all of the following elements:

mu1

p x r matrix of posterior means for the regression coeffcients.

S1

r x r x p array of posterior covariances for the regression coeffcients.

w1

p x K matrix of posterior assignment probabilities to the mixture components.

V

r x r residual covariance matrix

w0

K-vector with (updated, if update_w0=TRUE) prior mixture weights, each associated with the respective covariance matrix in S0

.

S0

r x r x K array of prior covariance matrices on the regression coefficients

.

intercept

r-vector containing posterior mean estimate of the intercept, if X_colmeans and Y_colmeans are provided. Otherwise, NA is output.

ELBO

Evidence Lower Bound (ELBO) at last iteration.

progress

A data frame including information regarding convergence criteria at each iteration.

converged

TRUE or FALSE, indicating whether the optimization algorithm converged to a solution within the chosen tolerance level.

Y

n x r matrix of responses at last iteration (only relevant when missing values are present in the input Y).

Examples

###Set seed
set.seed(123)

###Simulate X and Y
##Set parameters
n  <- 1000
p <- 100
p_causal <- 20
r <- 5

###Simulate data
out <- simulate_mr_mash_data(n, p, p_causal, r, pve=0.5, B_cor=1,
                             B_scale=1, X_cor=0, X_scale=1, V_cor=0)
#> Error in simulate_mr_mash_data(n, p, p_causal, r, pve = 0.5, B_cor = 1,     B_scale = 1, X_cor = 0, X_scale = 1, V_cor = 0): seed argument must be provided.

###Split the data in training and test sets
Ytrain <- out$Y[-c(1:200), ]
#> Error in eval(expr, envir, enclos): object 'out' not found
Xtrain <- out$X[-c(1:200), ]
#> Error in eval(expr, envir, enclos): object 'out' not found
Ytest <- out$Y[c(1:200), ]
#> Error in eval(expr, envir, enclos): object 'out' not found
Xtest <- out$X[c(1:200), ]
#> Error in eval(expr, envir, enclos): object 'out' not found

###Specify the covariance matrices for the mixture-of-normals prior.
univ_sumstats <- compute_univariate_sumstats(Xtrain, Ytrain,
                   standardize=TRUE, standardize.response=FALSE)
#> Error in ncol(Y): object 'Ytrain' not found
grid <- autoselect.mixsd(univ_sumstats, mult=sqrt(2))^2
#> Error in autoselect.mixsd(univ_sumstats, mult = sqrt(2)): object 'univ_sumstats' not found
S0 <- compute_canonical_covs(ncol(Ytrain), singletons=TRUE,
                             hetgrid=c(0, 0.25, 0.5, 0.75, 1))
#> Error in ncol(Ytrain): object 'Ytrain' not found
S0 <- expand_covs(S0, grid, zeromat=TRUE)
#> Error in ncol(mats[[1]]): object 'S0' not found

###Fit mr.mash
covY <- cov(Ytrain)
#> Error in is.data.frame(x): object 'Ytrain' not found
corX <- cor(Xtrain)
#> Error in is.data.frame(x): object 'Xtrain' not found
n_train <- nrow(Ytrain)
#> Error in nrow(Ytrain): object 'Ytrain' not found
fit <- mr.mash.rss(Bhat=univ_sumstats$Bhat, Shat=univ_sumstats$Shat, S0=S0, 
                   covY=covY, R=corX, n=n_train, V=covY, update_V=TRUE,
                   X_colmeans=colMeans(Xtrain), Y_colmeans=colMeans(Ytrain))
#> Processing the inputs... 
#> Error in mr.mash.rss(Bhat = univ_sumstats$Bhat, Shat = univ_sumstats$Shat,     S0 = S0, covY = covY, R = corX, n = n_train, V = covY, update_V = TRUE,     X_colmeans = colMeans(Xtrain), Y_colmeans = colMeans(Ytrain)): object 'univ_sumstats' not found

# Predict the multivariate outcomes in the test set using the fitted model.
Ytest_est <- predict(fit,Xtest)
#> Error in predict(fit, Xtest): object 'fit' not found
plot(Ytest_est,Ytest,pch = 20,col = "darkblue",xlab = "true",
     ylab = "predicted")
#> Error in plot(Ytest_est, Ytest, pch = 20, col = "darkblue", xlab = "true",     ylab = "predicted"): object 'Ytest_est' not found
abline(a = 0,b = 1,col = "magenta",lty = "dotted")
#> Error in int_abline(a = a, b = b, h = h, v = v, untf = untf, ...): plot.new has not been called yet